Midi tests to address dropped notes (#335)
* feat: log when note event queue is full

* test: Refactor MIDI test utilities and add MessageRouter tests

* refactor: split logger into different implementations

* refactor: replace midi_output_queue with midi_sender in drum

* fix: uncomment MIDI output queue processing in main loop

* style: Fix include order and formatting in main.cpp

* feat: use direct MIDI bypass for testing

* feat: add logging to MIDI output queue for diagnostics

* fix: Update logger calls to use correct log level and format

* style: Format code with clang-format

* fix: Pass logger to enqueue_midi_message calls in midi_wrapper.cpp

* style: Reorder includes in midi_wrapper.cpp

* feat: Add MidiSender class for flexible MIDI message sending

* fix: Pass logger to enqueue_midi_message calls

* feat: Add missing overloads for logger helper methods

* fix: Resolve ambiguous logger.error calls with explicit casts

* fix: Resolve vtable undefined reference for NullLogger

* fix: Coalesce MIDI CC messages and protect queue with spinlock

* style: Apply linter fixes

* fix: Replace etl::queue with etl::deque for iteration support

* style: format

* chore: Temporarily disable drum tests

* style: format midi sender

* fix: fix failing midi queue tests

* chore: no longer track ableton file

* chore: remove jsmn

---------

+++ drum/sysex/protocol.h
@@ -134,7 +134,7 @@ template <typename FileOperations> struct Protocol {
         return *maybe_result;
       }
 
-      logger.error("SysEx: Unknown command with no body. Tag", tag);
+      logger.error("SysEx: Unknown command with no body. Tag", static_cast<uint32_t>(tag));
       if (state == State::FileTransfer) {
         opened_file.reset();
         state = State::Idle;
@@ -236,7 +236,7 @@ private:
     case BeginFileWrite:
       return handle_begin_file_write(bytes, send_reply, now);
     default:
-      logger.error("SysEx: Unknown tag with body", tag);
+      logger.error("SysEx: Unknown tag with body", static_cast<uint32_t>(tag));
       if (state == State::FileTransfer) {
         opened_file.reset();
         state = State::Idle;
